[FIX] l10n_mx_edi: wrong cfdi date
Before this commit, the date send on the cfdi xml was the date and time
of  the invoice post, and not the date of the invoice date. This will
allow creating old invoices and sign it, and normally you cannot sign
invoices older than 72 hours.

Now, the invoice date is sent on the cfdi xml. Note that, this was
already the behaviour on 13. Also, as the invoice date don't have time,
we use the time of the invoice post.

opw-2475948

closes odoo/enterprise#17130

X-original-commit: 1b5fe20d0d646303fb06fc60bf3d80f26804c2f3
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>