[FW][FIX] crm_helpdesk: correctly take partner phone when converting lead
If we convert a lead without phone and mobile and assign newly
created ticket to a new partner its phone is not taken into
account. Indeed original code takes as granted that partner
is linked to the lead which is not always the case.

Task ID-2461308
Linked to COM PR odoo/odoo#67859

X-Original-Commit odoo/enterprise@074cd4616e9380a76ee5a17043219f83bb3bdbdf

closes odoo/enterprise#17191

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>