[FIX] helpdesk: Show Helpdesk overview tickets on mobile
A bug was introduced in the following PR: https://github.com/odoo/enterprise/pull/16020
where we tried to fix the display when we use "Group By" in
the Helpdesk Overview.

But on mobile, users weren't able to click on "Tickets" From the
dashboard...

To fix this we now use the right property: "sticky".

opw-2479945

closes odoo/enterprise#17111

closes odoo/enterprise#17173

X-original-commit: b1850b82f5ddc784cdcbc18d0ee91cb4f4a019cf
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>