[FIX] account_reports: don't crash when using action_id in account.report's lines dict on a report not using the generic id
saas-14.2 introduced a set of helpers to handle ids of account.report lines in a generic way. However, not all reports use them. Tax report, for example, still uses manual ids. Before this fix, such report crashed when trying to assign an action to a line, because their id was directly the model id (an int), not a generic id (a formatted string). An example of such use case can be found in l10n_be_reports, with the action added for lines 61 and 62 of the tax report (for tax adjustements)

closes odoo/enterprise#17112

X-original-commit: 659614c0cbd5c7f68eff2e5358b0b8bace1a3910
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>