[FIX] web_enterprise: fix m2o quickedit in mobile
Before this commit, clicking on the label of a many2one field triggered
the field's quick edit and so tried to open an autocomplete.

This commit patches the `_toggleAutoComplete` to call
`_searchCreatePopup` instead of opening an autocomplete.
It also removes the patch of `_onInputClick` as the non patched version
will do the same thing.

closes odoo/enterprise#17160

X-original-commit: b94750984b9695316dfcd74b42b0697da70ad37e
Signed-off-by: rfr-odoo <rfr-odoo@users.noreply.github.com>