[FIX] account_online_synchronization: fix multicurrency default value
When linking an online account with an existing journal, the currency
in the wizard was not adapted to the currency of the journal. Which, if
the user is not looking, will try to change the currency of his
existing journal. We change that behavior so that now when selecting
a journal, it will change the currency to match the one from the journal

Another problem is when we do not change the currency of the journal but
that journal contains entries in multiple currencies. It will results
in an error because there is a simple check that check the value of
the currency passed in the write method and does not verify that the
value has changed from before. To prevent that, we do not write
the value of the currency_id on the journal if it has not changed.

closes odoo/enterprise#17150

X-original-commit: 0bdb1ef9338728e3ec6f1bd3218f722b1e5ca5f7
Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>