[FIX] l10n_mx_edI: super called in return method
In the next case:
1. Generate a customer invoice
2. Confirm and stamp with the SAT
3. Request EDI cancellation

In Mexico, in the step 3, is necessary set the CFDI origin field with the UUID
from the invoice, to ensure that will to relate in the new invoice (If will be
validated again).

But in some cases (before of this change), the CFDI origin is not set because
the fiscal document is removed in the super, and the UUID is not found to be
saved on the invoice. For this reason, first try to set the CFDI origin and after
calls the super to set in draft the customer invoice.

closes odoo/enterprise#17219

X-original-commit: 1fbcc4eb2d4bd1b3083257b4860df1ec8ec3c983
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>