[IMP] website_calendar: schedule appointment link
This commits introduces custom and work hours appointment types.
The goal is to ease the scheduling of particular appointment type.
This feature is only enable if the user has an employee linked to it.

The first one represents custom appointment type which contains
real time slots. This slots are created from the calendar view of the
calendar app for the actual user.

A work hours appointment type is an appointment type that is used to
propose a link to select a slot from the working hours of an employee.

This appointment category is managed by a new field: "category"

* Calendar view:
	- Add a dropdown which display several options for the user
(creation of temporary appointment/slots and appointment types of the actual
employee).
	- Add a mode in the calendar view, to allow to create events as
before (default) and a mode to create custom slots for a custom appointment
type.
	- The end process of all the options allows to copy automatically the
link of the appointment type.

* Appointments Models:
	- Add one selection field for the category of an appointment type
	- Add fields to the slot model to manage real time slots

task-2287846
COM PR: odoo/odoo#69484
ENT PR: odoo/enterprise#17792
UPG PR: odoo/upgrade#2401

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>