[FIX] account_bank_statement_import_camt: journal with other currency
If the company's currency is C, when trying to import a CAMT file to a
journal with a currency defined but different from C, the new bank
statement will not contain any line.

To reproduce the error:
(Company's currency must be USD)
1. In Settings, enable:
    - Multi-Currencies
    - CAMT Import
2. In Accounting, create a new journal J:
    - Type: Bank
    - Currency: EUR
3. Import statements into J using this file:
    - [camt_053_minimal_EUR.xml](https://drive.google.com/file/d/1PyjqGgDeM12KVgo0Iux3FOxrM4ASIo35/view?usp=sharing)

Error: When confirming the import, the page is redirected to the
reconciliation page. However, there is nothing to reconcile. When
consulting the statement in J, the imported statement is present.
However, there is only one line in the statement, and this bank
statement line is zero (should be 500€)

Until now, the module reads the CAMT file and looks for an amount
expressed in the company's currency. However, it is not possible to find
an amount in the company's currency (USD) because all the amounts in the
CAMT file are expressed in journal's currency (i.e, EUR). This is the
reason why the bank statement line is zero.

Note: the file `camt_053_minimal_and_multicurrency.xml` has been added
to test #16656

OPW-2486849

closes odoo/enterprise#17515

X-original-commit: 77d236174226aa26ba6c9654526276000c9625d3
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>