[IMP] planning: increase nb of shifts displayed in the frontend
Currently in the planning, When the user always plans a shift in the
morning, and one in the afternoon, the employee won't be able to see
any of his shifts in the calendar view in the frontend because the
number of shifts displayed is limited to one.

So in this commit, changed eventlimit to 2 for daygrid of calendar
front end view. So employee can see  at least 2 shift for a day and
if an employee having more than two shifts on the same day then it
will show 2 shift and a other shift show in "number +more" link to
see all the shifts.

closes odoo/enterprise#17245

Task-id: 2463013
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>