[FIX] web_enterprise: properly clean up after tests
The PromoteStudioDialog blocks UI when the user clicks on "Install
Studio". The UI is then automatically unblocked as the webclient
is reloaded.

In tests, we mock the reload, but we didn't unblock the UI (i.e.
the class o_ui_blocked remained on the body). This is a side-effect
and we don't want tests to have side-effects, as it might influence
other tests. In this case, a web_tour test failed if it was
executed after those tests.

closes odoo/enterprise#17224

Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>