[FIX] snailmail_account_followup: fix trailing white space in acls file.
Since the changes from linked Community PR (see below), the import of boolean
is not assuming anymore that the value is "Yes" if a wrong value is given to the
import. For now on, the wrong value will raise an error.

In the ACLS file of this module, a trailing space was present in the last row of
the file. In the previous implementation of boolean import, that "0 " was
replaced by 'true' and was then misleading to allow the deletion of records for
the given group.

since the new bollean import implementation, the "0 " value crashed while
importing.

This commit removes that trailing space to correctly set the access rights to
the given user group.

Linked COM PR: odoo/odoo#61948
Task ID: 2352241

closes odoo/enterprise#17246

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>