[FIX] industry_fsm_sale: fix required value for timesheet product
Currently, when go to project > edit 'office design' > change its name
> save. the user is blocked because the timesheet product is not set.

So in this commit, Make the timesheet product required based on
condition as it's not always required. now the timesheet_product_id
field is only required for fsm project which is billable and allows
timesheets.

Also the 'invoicing' tab should not be visible if allow_billable is
false and is_fsm is true as the timesheet_product_id only required
if billable project.

closes odoo/enterprise#17625

Taskid: 2459782
X-original-commit: cbb5a173479dd38825ad1d6984c556a7ebe54ed6
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Mahendra Barad <mba-odoo@users.noreply.github.com>