[IMP] hr_contract_salary_holidays: link allocation to contract
Link automatically created time off allocation to their contracts.
Add a link from the contract form view to the time off allocation view.
A linked time off allocation will be cancelled if the contract is
cancelled.

Change was made to reduce steps to remove an allocation in the case of a
contract needing to be signed again (new contract).

Task ID#2487397

closes odoo/enterprise#17271

Related: odoo/upgrade#2367
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>