[FIX] account_online_synchronization: post stmt line automatically
Statement lines added to an existing posted statement should have
their move posted automatically because, after adding new transactions,
we open the reconciliation widget with those newly added lines and it
is not possible to reconcile a statement line which has its associated
move in draft.

Automatically posting the newly fetched transactions means that a
sequence number will be taken even if the transaction is a duplicate
but we consider that it is not much of a problem as we can have
holes in sequence number within the bank journal.

closes odoo/enterprise#17276

X-original-commit: 4a73eddb9c2909c2a61e23698f4a98a0f9f739b9
Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>