[FIX] voip: Avoid double dialing panel display on mobile devices
When opened on a mobile device (smartphone or tablet), both the native
Dialing Panel and the virtual one opens when composing a phone number.

This commit sets the DialingPanel Widget's flag to detect such
environment and properly hides the virtual dialing panel when the native
one is available.

Steps to reproduce:

1. Open VOIP panel
2. Click on dialing panel to compose a phone number
3. Both dialing panels are displayed => bug

Task ID: 2495658

closes odoo/enterprise#17430

X-original-commit: 4c0a14f932bdc3c6acf311d5d959529f665cbef7
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>
Signed-off-by: res-odoo <res-odoo@users.noreply.github.com>