[FIX] l10n_ar_edi: avoid AFIP accesses in test mode
In test mode, we should not access to AFIP web service to get
the last invoice number as the AFIP key may be not configured.
Instead of that, we should return a default value.

One reason is that during an upgrade, we check that all the menus/views have
the same behaviour before and after the upgrade to check that not menus/views
have been broken. During this check, when we access to the `menu account.menu_action_move_out_invoice_type`,
`account.move.highest_name` is computed and, in its compute method, `account_journal._get_last_sequence_from_afip()`
 is called, which leads to an error as the AFIP key is not configured.

The test mode is enabled only when `self.registry.enter_test_mode(cr)` is
explicitely called. This is the case for upgrade tests for example, but not for
`l10n_ar_edi` tests.

upg-4898
opw-2481780

closes odoo/enterprise#17321

X-original-commit: 8e47e74675ff563f9017b0b61bfd485c0b7904e9
Signed-off-by: Josse Colpaert <jco@openerp.com>