[IMP] account_reports: tax report: allow opening of foreign reports if needed
A foreign tax report is considered available (and hence: is selectable in the report's menu) if there exists at least one foreign VAT fiscal position for its country in the company.

[IMP] account_reports: introduce fiscal_position option

This options is used to filter what is displayed in the report depending on the foreign VAT fiscal positions. Available fiscal positions are obtained by checking the country of the report (the way this check is performed depends on the report type, but it's available for all of them, no only account.generic.tax.report).

A 'domestic' and 'everything' option also exist, respectively to only show domestic operations (i.e. the ones without any fiscal_position_id), or all of them (domestic + all other fiscal positions from this country). Note that 'domestic' will not be available in case the country defined for the report does not match the company's fiscal country.

In case a report is opened for a country that's not domestic and contains only one foreign VAT fiscal position, the filter is not displayed in the report, and this fiscal position is always selected. So, in most common cases, the option will stay hidden in the interface. It will only show if there are multiple VAT fiscal positions within the same country (for example, if a different tax report needs to be submitted per region, like in India).

[IMP] account_reports: tax closing: one closing move per VAT fiscal position and tax report

There is not a single tax closing move anymore. It's now unique per (country, fiscal_position). So, you can have multiple closing moves for the same period at the same time, for different operations in different areas. fiscal_position_id of these moves contain the foreign VAT fiscal position that was used to generated them, or None if the move is the closing for the domestic region.

[IMP] account_reports, l10n*: add country_id to financial reports

It is now possible to define which country a report belongs to. If not set, this field indicates that the report is not specific to a single country.