[FIX] account_accountant: reconcile only posted lines
The reconcilation page also suggests to reconcile the bank statement
lines not posted yet.

To reproduce the error:
1. Create a bank statement with one line BSL
2. Back to Accounting Dashboard, observe the Bank's Kanban Card
    - There is a button "Reconcile X Items"
3. Click on this button

Error: There are actually X+1 items to reconcile. The difference comes
from BSL. Moreover, since BSL is not posted, the user can not reconcile
it ("You can only reconcile posted entries."). Therefore, X is the
correct value and BSL should not be listed while it is not posted.

OPW-2469365

closes odoo/enterprise#17440

X-original-commit: be1b155129b95fe9065cee70e633a7f6b7f398ec
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>