[FIX] pos_blackbox_be: automatically list modules contents
Currently the list of files to report on is hard-coded, meaning it
easily gets out of sync as files get removed (at least two of the
listed files don't exist anymore) or added back in (4 of the JS files
from pos_blackbox_be are unlisted).

Automatically enumerate all interesting files (= python, js, views,
templates, and data), in lexicographic order (so the hash can stay
stable), across the two relevant modules.

Use pathlib as it seems to be less of a mess than os.path here.

closes odoo/enterprise#17337

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>