[IMP] web_mobile: adapt mobile PivotRenderer
The groupby selection in a pivot view is now done via a menu similar to
the "Group By" menu in control panel: the groupbys proposed when a user
click on a suitable header are those defined in the search arch
and a "Add custom group" menu is also available for more flexibility.
For more information and precision see https://github.com/odoo/odoo/pull/61572.
Here we simply adapt the mobile PivotRenderer to the above mentionned
changes.

Task ID: 2376505

closes odoo/enterprise#17341

Related: odoo/odoo#61572
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>