[FIX] l10n_be_hr_payroll: avoid negative wage with unpaid hours
Unpaid hours are removed from the wage. A hourly wage is computed to
deduct the same amount in small months (february) and large months (march),
by knowing that there is always 13 weeks over 3 months, the hourly wage can
be computed as:

hourly_wage = wage * 3 / (13 * hours_per_week)

If there is only 1 day of paid remuneration over a large month like March,
this could lead to negative amount, so fix the minimal amount to 0.

X-original-commit: f794504d032e253cc56e4ffd162d8fd764e4e721