[IMP] project_forecast: review dropdown menus in project kanban card
Before this commit, we add differents buttons in the bottom of each card
with a certain condition, the problem is the lack of place to display
more and more information in each project kanban card.

In community,the dropdown menus are reviewed for each project kanban card.
Two sections is added one for 'Views' and the other for 'Reporting'.
This commit add a button called 'Planning' in the 'Views' section rather
than in the bottom of the project kanban card.

task-2458017

Community PR: odoo/odoo#68343

closes odoo/enterprise#17378

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>