[IMP] industry_fsm_sale: add unit tests for employee rate in fsm project
Before this commit, very few tests exist to test the behaviour in a fsm
project with pricing type set to employee rate. That is, with employee
mappings.

This commit adds some unit tests to check different cases that could be
arrived when a fsm project has employee mappings. That is, the SOL
generated in the SO generated to a fsm task when the user marks as done
the task. When the user changes the SOL linked to the task, we need to
search the SOLs for the timesheets based on the mappings, etc.

task-2376382

Related PR: #16348

closes odoo/enterprise#17413

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>