[FIX] web_studio: no block on error on xml editor
When editing XML view inside studio, if an error happened on save: the
save button was never unlocked and you needed to refresh and lose
current changes to continue editing.

With this changeset, the promise rejection is available in calling code
which can unblock the save button.

opw-2460081

closes odoo/enterprise#17395

X-original-commit: 100a961a97bfc284064d4c1d3d789ba98882546d
Related: odoo/odoo#68569
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>