[FIX] sign: reposition tooltip to avoid resize flicker
The tour bubble "animation" that makes it to bounce up and down can cause
issues when its position is at the edge of the bottom of the screen.

In the sale tour, this would make the window constantly resize to show a
scrollbar and then resize to hide the scrollbar, creating quite a sickening
effect visually.

closes odoo/enterprise#19147

Task-id: 2480195
X-original-commit: 9ddd11d3cb4e5d63e05a58fffe06821fcb2c0e90
Related: odoo/odoo#72485
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>