[FIX] Social marketing: avoid refreshing posts without a linkedin_post_id
In the social marketing app, when you click on the app the system will try to refresh statistics
from the Linkedin posts (social_live_post in DB with the associated social account). If one of
these post doesn't have a linkedin_post_id (eg. if the post is failed for this account), it
will trigger a TypeError.

Now we are filtering posts to only refresh the statistics on posts with a linkedin_post_id.

opw-2463073

closes odoo/enterprise#18533

X-original-commit: 8b7952e02451046be2d90c68e5749918d5fa1607
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>