[FIX] data_merge: support inherited tables
Steps:
- Install data_merge, product
- Go to Data Cleaning / Configuration / Rules / Deduplication
- Create a new rule:
  - Model: Product
  - Deduplication Rules:
    - Barcode (product.product)
- Deduplicate

Bug:
UserError: Missing required PostgreSQL extension: unaccent

Explanation:
Whenever the SQL query is not valid, a UserError is raised claiming that
the extension `unaccent` is missing.
This commit raises the original error if it has nothing to do with
missing this extension.

The original error was due to the fact that the model has a `company_id`
field, but it's inherited.
This commit uses the ORM to rework the query and add inherited tables.
This should work since we filter out the non-stored related fields:
https://github.com/odoo/enterprise/blob/74efaf093b44d224dd4fd4868de595269099b2c6/data_merge/models/data_merge_rule.py#L13-L15

opw:2476683

closes odoo/enterprise#17755

X-original-commit: 3be20f9d8b4b73f4c62adf6bfedf581d003e5efe
Signed-off-by: backspac <backspac@users.noreply.github.com>