[REF] *: adapt to legacy js files move
To prepare the rewriting of the webclient in owl, a community
commit moved all current js files of web in a legacy/ folder.
This folder will eventually be removed, as soon as each file
it contains will be converted to owl and moved to the proper place
in the new file structure.

This commit adapts the code in enterprise to those changes.