[IMP] social_*: introduce a computed "message" field on social.live.post
This commit introduces a new concept of computed "message" field on the
social.live.post model.

As a reminder, a social.live.post is a "decorated m2m" between social.post and
social.accounts.
Basically, a social.live.post is created for each account you want to publish
your content when posting a social.post.

This new "message" field allows to centralize some pre-processing that has to
be done before publishing the content on the social media.
For example, we want to shorten links and UTMize them.
(This also applies to displaying previews).

The main reason is also that we are going to need this new mechanism to
implement the YouTube social media.
Indeed, social.live.post content needs to be customized before posting to
contain the YouTube video link.
(Some references to YouTube are kept in docstring in this commit to avoid too
much splitting and give proper implementation reasons).

Task-2209729
ENT PR odoo/enterprise#14027