[FIX] industry_fsm: ensure partner_id is correctly set on task
Prior to this commit:
The partner_id was not set on non FSM projects tasks once the inductry_fsm
module was installed.
The partner_id field was not displayed in the kanban quick_create view as the
is_fsm field was Falsy since the project_id field was not present in the view.

Both behaviors have been introduced with odoo/odoo#66589.

After this commit:
The partner_id is correctly set as the compute is triggered thanks to project_id.
The partner_id field is visible in the task quick_create view for FSM projects since the is_fsm
task field works as expected as the project_id is present in the view.

closes odoo/enterprise#17437

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>