[FIX] mrp_workorder: close WO with remaining QP
Before this commit, it was possible to close a Work Order without pass
its Quality Checks with the help of the "Start"/"Done" buttons.
As these QC can only be proccessed through the WO tablet view, it wasn't
possible anymore to mark as done the WO's Manufacturing Order.

To avoid that, Odoo will no raises an UserError if the user tries to
mark as done a WO with remaining Quality Checks.

task-2496604

closes odoo/enterprise#18508

X-original-commit: 2adc99f1d2af473a77b30270704caa960de36d62
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>
Signed-off-by: Steve Van Essche <svs-odoo@users.noreply.github.com>