[IMP] sign: adds edit mode while signing document
Adds edit mode during sign procedure, allowing back-end users to add fields to
a template while signing a document. The rationale behind this feature is: when
you fill a whole document and see it miss a field, you have to quit the sign
process, ask for edit, start from beginning again. The process is quite
frustrating if you use sign often.

This PR adds an extra flow to the sign application, allowing internal users to
edit already sent requests while signing. By using ctrl+click, the user is now
able to add new sign items (only signature, initials and text types) to a
request if the request has not been signed by anyone yet.

Technically, the user adds the new items to a new template that is a copy from
the original one. This process adds an "Update" log to the sign request saving
who edited the template and when it was done. Since the changes are done to a
new template and with the guarantee that no one has signed the document yet,
the security of sign is not changed by this feature. In order to guarantee this,
automated tests were created to assure the feature is working correctly.

Two options were available to allow this feature:

First, a button could have been added on the control panel of the backend view.
It would have redirected to the template edition view. This solution would be
simplier to implement but it would have broken the flow and it may have been
difficult to keep the item already filled when switching from the signing to
editing views.

The other option was to directly add items during the signing procedure and send
all the items (old and new) to the sign controller to handle them properly. This
solution has been kept as the flow is less perturbed.

Furthermore, it was decided to keep an uncommon flow when one can send a template
without item. In that case, the "Sign Document" button in the control panel allows
to validate it without use of any item. To make it work, a default role was needed
to allow the creation of item when no role was available. The default role is
therefore mandatory for this use case. That's why new ir.rule are added in this
commit to prevent deletion of them by regular or even admin users.

task-2475732

closes odoo/enterprise#17574

Related: odoo/upgrade#2590
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>