[FIX] l10n_ar_edi:Set the context to return the datas when getting the AFIP private key and certificate
Getting AFIP private key and the certificate can be called with a context containing
"bin_size" at True which will return the size of the attachment and not his datas
(eg. try to open a debit note that call a "_compute_highest_name") The system will
try to decode the size returned as if it was the datas and result to an error.

So bin_size=False is added to the context to always get the datas needed.

opw-2453951

closes odoo/enterprise#17486

X-original-commit: 472fad2b58303ac5ac9252117ada095872d2cbbe
Signed-off-by: Josse Colpaert <jco@openerp.com>