[FIX] mrp_account_enterprise: fix scrap value
In the cost structure report, the scrap valuation is computed from the
stock valuation layer price unit. But the value is computed from this
price unit and the scrap quantity, not the valuation layer value itself.
With rounding errors correction introduced by 74ba8039c7e300bb2adeda8cd071df4272b3635a
the value can be different than the quantity multiplied by the rounded
price_unit.

Not that the raw material already take the value directly on the layer
to render the cost structure report.

opw :2474923

closes odoo/enterprise#17485

X-original-commit: 7397da633f1503f2cfe2cc50681f36b7394270a2
Signed-off-by: Rémy Voet <ryv-odoo@users.noreply.github.com>
Signed-off-by: William Henrotin <Whenrow@users.noreply.github.com>