[REF] web_studio: move legacy files
To prepare the rewriting of the webclient and studio in owl, we
move all current js files in a legacy/ folder. This folder will
eventually be removed, as soon as each file it contains will be
converted to owl and moved to the proper place in the new file
structure.