[IMP] social_push_notifications: add an event listener to trigger a new push notification request
PURPOSE

Currently, Odoo will ask the user to enable push notifications when the
user accesses the website for the first time. If the user dismisses the
request, no other request will be issued within the 7 upcoming days.

When the user adds a track to his/her favorites, we really want to make
sure that the push notifications are enabled so that the user can be
notified when the track is about to start. If the user instinctively
dismisses the request, we will not be able to issue a new one immediately.

With these new changes, the module having the responsibility of issuing
a new push notification request will listen to a new event called
`open_notification_request`. The programmer will then be able to trigger
a new request by emitting that new event. When the module intercepts that
event, it will issue the request if no other request for the same provided
key has been issued within the last 7 days.

With this new system, we can trigger a new push notification request when
the user adds a track to his/her favorites even though the user dismissed
the first notification request. By propagating the event with the same key,
we can ensure that we will not keep opening popups if the user does not
want to enable push notifications.

SPECIFICATIONS

- Add an event listener to trigger a new push notification request.
- Ensuring that there will be at most one push notification request at a time.

LINKS

Task-2347597
COM PR odoo/odoo#69102
ENT PR odoo/enterprise#17612

X-original-commit: 8580348f92fba234d87b67d98cc6dd77e0f7efeb