[FIX] account_asset: comparing disposal_date with itself when there are no past posted depreciation moves
If it happened that disposal date is in the past and there are no posted depreciation moves that are less than the disposal date the disposal date get compared with today's date which results in entering the if block, thus get an error for no legit reason.
The bug as I encoutered it, I was not able to sell an asset with an invoice from the past.
opw-2476113

closes odoo/enterprise#17533

X-original-commit: d47c4767b232b6238426c73d06b4d6b8d1f7b506
Signed-off-by: William André (wan) <wan@odoo.com>