[IMP] hr_appraisal: Improve main appraisal flows (back2basics)
This commit consists of the following major changes:

Appraisal:

When the appraisal is requested from an employee form, or from My Profile,
the wizard does not pop up anymore (which used to create an appraisal).
Now, the button redirects to an appraisal form.

In the appraisal form ‘SEND BY MAIL’ button is added.
The button simply sends an appraisal request.
When the email is sent, the copy is added to the chatter.
Activity is not created anymore.

Appraisal rights are updated.
Now appraisal rights do not involve any more rights on Employees.
Also, appraisal rights do not involve any more rights on Surveys.
Rights are granted to base.group_user users that are implicit managers.
A base user A is an implicit manager for an employee B under two circumstances:
- If A is a manager of B (In employee app).
- If A has been one of the managers of B on one of B's appraisals (historical reason).

Now appraisal is linked to 'hr.department' instead of ‘hr.job'
- ‘Custom Appraisal Templates’ field is added, which allows to
customize appraisal templates by department. If ‘custom appraisal template’ is
checked in the employee’s department - take it, else take them from the company.

Major changes in Appraisal form
- Note tab is added.
- For an employee or manager, before assessment is published it takes up full tab.
 After it is published, it takes up only a half and the other column displays
 the other persons assessment. Blurred if not published.
- When the appraisal is marked as done, both feedbacks are published.

Change an appraisal filter ‘To Do’ and add ‘Archieved appraisals’ filter.
Display the count of not completed goals linked to the Employee.

Skills

Skills are added to the appraisal
- The skills are copied from Employee app when an appraisal is confirmed.
- skills can be modified by an appraisal manager. New skills can be
added and existing can be deleted.
- On top of the other fields, hr_appraisal_skills has ‘justification’ field.
- When the appraisal is marked as done, the skills are updated in employee app.
- The changes in skills are logged in the appraisal chatter.

Goals

- When manager creates a goal, same logic applies as when creating an appraisal.
	Specifically:
	Appraisal manager can create a goal for anybody.
	Implicit (simple) manager can create a goal for himself and his subordinates.
- UX changes kanban and list view.

Hr.apprasial.plan is removed and scheduling of the future appraisals is moved to res.company.

UX changes in Emails

- Request appraisal mail is modified.
- When the appraisal is confirmed, the same email is sent for manager and employee.
- UX changes in ‘appraisal feedback’ email.

task - 2456977

closes odoo/enterprise#16774

Related: odoo/upgrade#2407
Signed-off-by: Kevin Baptiste <kba@odoo.com>