[IMP] project_forecast: get assigned user to the task in planning slot
Before this commit, when the planning slot has no employee and the user
adds a task in the slot, this slot has always no employee and the user
must to manually define the employee of the slot.

This commit gives by default the employee_id of assigned user to the
slot when the user adds a task into this slot and this slot has no
employee.

task-2463722

Upgrade PR: odoo/upgrade#2426

closes #17958