[IMP] planning: update planning.slot gantt view
Before this commit, when the user clicks on 'Add' button to create a new
planning.slot, the default dates are based on the scale of the gantt
view but most of the time the user wants to create a slot for one day.
Moreover, when the user clicks on the 'Send Schedule' button and there
are no shift in the gantt view, the user can fill the wizard and valid
to have an error message saying there is no shift to send.

This commit changes the default dates to have a new shift on one day
(we use the focus date if the user is in the current week/month else we
take the first workday) and add a toast notification when there is no
shift to send rather than open the wizard after clicking to the
'Send Schedule' button.

task-2463722

Upgrade PR: odoo/upgrade#2426

closes #17958