[IMP] planning: use default value for repeat_interval and repeat_type
Before this commit, the computes for repeat_interval and repeat_type
fields do not use the default value of these both value when the
recurrence_id is False. This commit removes to falsify the both fields
when the recurrence_id to use the default value when no value is set for
these fields in their compute method.

Upgrade PR: odoo/upgrade#2426

closes #17958