[FIX] documents: fix many2one creation
Before this commit:

In documents, when you upload a document, if you want to create new record
like contact, owner or workspace in the right panel, you write it and click on
"create", but the field is not loaded and you don't find it in the list also.

After this commit:

Records are created after clicking on create, added to list and will be
selected.

Task-2487738

closes odoo/enterprise#27703

X-original-commit: f2a8e49e00f0dc05e8fd6ff697c7df8a3b20a506
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>