[FIX] web_enterprise: 2nd x2many full-width if 1st invisible
In eg. stock picking view, we want to display one out of two x2many but
only the first one was shown without margin.

So the page was duplicated so the x2many is always first of its page,
but this duplication can cause problems eg. for studio customization
that currently have for fields a selector based on the field name or
subviews.

With this change, we add style so if the first field of a page is
invisible, we show a second x2many without margin instead

=> this allows us to remove the code duplication (at least in master)

opw-2480311

closes odoo/enterprise#17571

X-original-commit: 622f0e1ce2f21538c91f60085a19639a706219f5
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>