[FIX] l10n_ar_edi: Correct the context to return the datas when getting the AFIP certificate or private key
Getting AFIP the certificate or private key can be called with a context containing
"bin_size" at True which will return the size of the attachment and not his datas
(eg. try to open a debit note that call a "_compute_highest_name") The system will
try to decode the size returned as if it was the datas and result to an error.

This completes the commit 472fad2b58303ac5ac9252117ada095872d2cbbe

opw-2453951

closes odoo/enterprise#17748

X-original-commit: 03032e7a58292d17cbe750d0a5a609daffb47e10
Signed-off-by: Josse Colpaert <jco@openerp.com>